feat: Add screen functions for displaying alarm icons
The code changes in this commit add screen functions for displaying alarm icons on the alarm clock interface. The `screen_repeat_alarm_icon()` function is added to display a repeat alarm icon when the alarm is set to repeat. The `screen_active_alarm_icon()` function is added to display an active alarm icon when the alarm is currently ringing. These functions use predefined screen coordinates to position the icons on the screen.

Note: This commit message is generated based on the code changes and recent repository commits.